% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signaling2020}
\alias{signaling2020}
\title{signaling2020}
\format{
A data frame with 1950 rows and 61 variables:
\describe{
\item{\code{MTurk ID}}{character COLUMN_DESCRIPTION}
\item{\code{IPAddress}}{character COLUMN_DESCRIPTION}
\item{\code{Progress}}{double COLUMN_DESCRIPTION}
\item{\code{sample}}{character COLUMN_DESCRIPTION}
\item{\code{vignette}}{character COLUMN_DESCRIPTION}
\item{\code{signal}}{character COLUMN_DESCRIPTION}
\item{\code{T1Belief}}{double COLUMN_DESCRIPTION}
\item{\code{T1Action}}{double COLUMN_DESCRIPTION}
\item{\code{T1Divide}}{double COLUMN_DESCRIPTION}
\item{\code{T2Belief}}{double COLUMN_DESCRIPTION}
\item{\code{T2Action}}{double COLUMN_DESCRIPTION}
\item{\code{T2Divide}}{double COLUMN_DESCRIPTION}
\item{\code{T3Action}}{double COLUMN_DESCRIPTION}
\item{\code{T1Angry}}{double COLUMN_DESCRIPTION}
\item{\code{T1Sad}}{double COLUMN_DESCRIPTION}
\item{\code{T1Suicidal}}{double COLUMN_DESCRIPTION}
\item{\code{T1MentallyIll}}{double COLUMN_DESCRIPTION}
\item{\code{T1Depressed}}{double COLUMN_DESCRIPTION}
\item{\code{T1Guilty}}{double COLUMN_DESCRIPTION}
\item{\code{T1Calm}}{double COLUMN_DESCRIPTION}
\item{\code{T1Neutral}}{double COLUMN_DESCRIPTION}
\item{\code{T1Scared}}{double COLUMN_DESCRIPTION}
\item{\code{T1Tired}}{double COLUMN_DESCRIPTION}
\item{\code{T1Distressed}}{double COLUMN_DESCRIPTION}
\item{\code{T1Devious}}{double COLUMN_DESCRIPTION}
\item{\code{T1Jealous}}{double COLUMN_DESCRIPTION}
\item{\code{T1Confident}}{double COLUMN_DESCRIPTION}
\item{\code{T1Traumatized}}{double COLUMN_DESCRIPTION}
\item{\code{T1Violated}}{double COLUMN_DESCRIPTION}
\item{\code{T1NoneOfAbove}}{double COLUMN_DESCRIPTION}
\item{\code{T2Angry}}{double COLUMN_DESCRIPTION}
\item{\code{T2Sad}}{double COLUMN_DESCRIPTION}
\item{\code{T2Suicidal}}{double COLUMN_DESCRIPTION}
\item{\code{T2MentallyIll}}{double COLUMN_DESCRIPTION}
\item{\code{T2Depressed}}{double COLUMN_DESCRIPTION}
\item{\code{T2Guilty}}{double COLUMN_DESCRIPTION}
\item{\code{T2Calm}}{double COLUMN_DESCRIPTION}
\item{\code{T2Neutral}}{double COLUMN_DESCRIPTION}
\item{\code{T2Scared}}{double COLUMN_DESCRIPTION}
\item{\code{T2Tired}}{double COLUMN_DESCRIPTION}
\item{\code{T2Distressed}}{double COLUMN_DESCRIPTION}
\item{\code{T2Devious}}{double COLUMN_DESCRIPTION}
\item{\code{T2Jealous}}{double COLUMN_DESCRIPTION}
\item{\code{T2Confident}}{double COLUMN_DESCRIPTION}
\item{\code{T2Traumatized}}{double COLUMN_DESCRIPTION}
\item{\code{T2Violated}}{double COLUMN_DESCRIPTION}
\item{\code{T2NoneOfAbove}}{double COLUMN_DESCRIPTION}
\item{\code{Age}}{double COLUMN_DESCRIPTION}
\item{\code{Sex}}{character COLUMN_DESCRIPTION}
\item{\code{Sons}}{double COLUMN_DESCRIPTION}
\item{\code{Daughters}}{double COLUMN_DESCRIPTION}
\item{\code{Rel Status}}{character COLUMN_DESCRIPTION}
\item{\code{Ed}}{character COLUMN_DESCRIPTION}
\item{\code{Income}}{character COLUMN_DESCRIPTION}
\item{\code{Feedback}}{character COLUMN_DESCRIPTION}
\item{\code{TotalTime}}{double COLUMN_DESCRIPTION}
\item{\code{VignetteTime}}{double COLUMN_DESCRIPTION}
\item{\code{SignalTime}}{double COLUMN_DESCRIPTION}
\item{\code{T4AttentionCheck}}{double COLUMN_DESCRIPTION}
\item{\code{T4AttentionCheckFail}}{double COLUMN_DESCRIPTION}
\item{\code{CompleteSurvey}}{logical COLUMN_DESCRIPTION}
}
}
\usage{
signaling2020
}
\description{
This package contains cleaned data for both the US and Indian Samples of our 2020 vignette study.
}
\details{
DETAILS
}
\keyword{datasets}
